refactor(inbox): extract the Files column (#994)
Moves the FR-011 mixed-composition summary, the FR-010 per-file metadata
popover (table + FileInspector) and the FR-032/US9 missing-attribute
banner into InboxFilesColumn.tsx, together with the metadata column/row
builders and the popover's inspected-row state — none of which anything
outside that column read.

InboxDetail.tsx: 627 -> 431 lines, completing the #994 split.
